fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -1,6 +1,12 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
+});
+
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) {
@@ -13,11 +19,14 @@ const uploadOnCloudinary = async (localFilePath) => {
         })
 
         // File has been uploaded successfully
-        console.log("File is uploaded on cloudinary", response.url);
+        // console.log("File is uploaded on cloudinary", response.url);
+        fs.unlinkSync(localFilePath)
         return response
         
     } catch (error) {
-        fs.unlinkSync(localFilePath) // Remove the locally saved temporary file as the upload operation got failed
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath)
+        }
         return null;
     }
 }
